+++ bin/compile.php
@@ -15,12 +15,5 @@ use BEAR\Skeleton\Injector;
 require dirname(__DIR__) . '/vendor/autoload.php';
 
 $context = $argv[1] ?? 'prod-app';
-$tmpDir = null; // ={appDir}/var/tmp/{context}
-$logDir = null; // ={appDir}/var/log/{context}
 
-exit(
-    Compiler::fromInjector(
-        Injector::getInstance($context, $tmpDir, $logDir),
-        $context,
-    )->run()
-);
+exit(Compiler::fromInjector(Injector::getInstance($context), $context)->run());